perf: port some perf improve to 80 branch (#12501)

+++ src/main.rs
@@ -697,8 +697,7 @@ async fn init_http_server() -> Result<(), anyhow::Error> {
             let handle = handle.clone();
             async move {
                 shutdown_signal().await;
-                handle
-                    .graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout as u64))));
+                handle.graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout))));
             }
         });
 
@@ -764,8 +763,7 @@ async fn init_http_server_without_tracing() -> Result<(), anyhow::Error> {
             let handle = handle.clone();
             async move {
                 shutdown_signal().await;
-                handle
-                    .graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout as u64))));
+                handle.graceful_shutdown(Some(Duration::from_secs(max(1, shutdown_timeout))));
             }
         });
